rotated labels tested all cases
