+++ Examples/Wpf/CartesianChart/Labels/LabelsExample.xaml
@@ -5,8 +5,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:Wpf.CartesianChart"
              xmlns:lvc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
+             xmlns:labels="clr-namespace:Wpf.CartesianChart.Labels"
              mc:Ignorable="d" 
-             d:DesignHeight="300" d:DesignWidth="300" d:DataContext="{d:DesignInstance local:LabelsExample}">
+             d:DesignHeight="300" d:DesignWidth="300" d:DataContext="{d:DesignInstance labels:LabelsExample}">
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto"></RowDefinition>
@@ -18,16 +19,17 @@
             Move All
         </Button>
         <lvc:CartesianChart Grid.Row="2" Series="{Binding SeriesCollection}" 
-                            DataClick="Chart_OnDataClick" >
+                            DataClick="Chart_OnDataClick"
+                            Background="#606060">
             <lvc:CartesianChart.AxisX>
-                <lvc:Axis LabelsRotateTransform="{Binding RotateTransform}" Labels="{Binding Labels}">
+                <lvc:Axis LabelsRotation="35" Labels="{Binding Labels}">
                     <lvc:Axis.Separator >
-                        <lvc:Separator IsEnabled="False" Step="1"></lvc:Separator>
+                        <lvc:Separator Step="1"></lvc:Separator>
                     </lvc:Axis.Separator>
                 </lvc:Axis>
             </lvc:CartesianChart.AxisX>
             <lvc:CartesianChart.AxisY>
-                <lvc:Axis Title="Sold Items" LabelFormatter="{Binding Formatter}"></lvc:Axis>
+                <lvc:Axis Title="Sold Items" LabelFormatter="{Binding Formatter}" LabelsRotation="-35" Position="RightTop"></lvc:Axis>
             </lvc:CartesianChart.AxisY>
         </lvc:CartesianChart>
     </Grid>
