+++ Examples/Wpf/CartesianChart/Labels/LabelsExample.xaml.cs
@@ -2,7 +2,6 @@
 using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Media;
 using LiveCharts;
 using LiveCharts.Defaults;
 using LiveCharts.Wpf;
@@ -37,7 +36,6 @@ namespace Wpf.CartesianChart.Labels
                 }
             };
 
-            RotateTransform = new RotateTransform(13);
             Labels = new[]
             {
                 "Shea Ferriera",
@@ -55,7 +53,6 @@ namespace Wpf.CartesianChart.Labels
         }
 
         public SeriesCollection SeriesCollection { get; set; }
-        public RotateTransform RotateTransform { get; set; }
         public string[] Labels { get; set; }
         public Func<double, string> Formatter { get; set; }
